Fix issue #121104681: Force stopping AGSA causes system to revert to its default impl
If you have a legacy assist app (launched through ACTION_ASSIST, not
implementing a VoiceInteractionService) then we kept the Google app
as the current speech recognizer.  Thus if you force stopped the
Google app, we saw that as one of the active parts of the interactor
or recognizer and reset all of the state.

Now we handle the "only recognizer" case separately, only resettting
the recognizer state in that case.

Bug: 121104681
Test: manual
Change-Id: Icc007bdf2352548d58be997fae77d9e5aba842f3