Apply splash screen and transfer starting window to all activities
This fixes a CTS failure and lack of splash screen for some activities.

When removing the global lock in AppWindowToken in change e5907c4,
transferStartingWindow and creating splash screen was accidentally
wrapped by if(theme != 0) by mistake. The error will prevent simple app
from showing splash screen.

This patch fixes the error and the function is now aligned with the old
AppWindowContainerController#addStartingWindow.

Test: atest ActivityMetricsLoggerTests
Test: Test app in b/120968158 will show the splash screen with correct
      log information.
Bug: 120968158
Bug: 120779200

Change-Id: I34bf01874eecc3c0402d83d922445cd48ca1b39a