OsuLogin: Do not load the page for local IP address
Currently the OsuLogin Weview shows "web page not available" page when
user clicks getPasspoint Button on the page.
This is because we are trying to load the page for http://127.0.0.1 on
last moment before we exit the OsuLogin webview as part of HTTP redirect
to local server.

So, it needs to stop loading the page for the local server.

Bug: 121281421
Test: Tested with R2 AP
Change-Id: I37e3aa5985d49d0a58e367ff221d53df371551e0
Signed-off-by: Ecco Park <eccopark@google.com>