Applies friction to non-dismiss drag only after the snap point.
This change modifies SwipeHelper such that the friction applied to a
constrained gesture (i.e. one that won't dismiss the swiped object) is
only applied once the object has been dragged past a threshold defined
by the callback object.  The callback provided by
NotificationStackScrollLayout is modified to set this threshold at the
snap point of the active menu row.

Change-Id: Ide9261682963d3cfaeb470e08393c182d6fa06c6
Fixes: 123021652
Test: atest SystemUITests, manual