Remove code that does nothing
Test: CtsGraphicsTestCases

When the comment was written, "the default" referred to
GraphicsJNI::colorSpaceForType, which was either SRGB or null, depending
on a build flag. In If7b1c00ad889446658f98e9ad644869e2eeabcdb, we
changed it to "decodeInfo.refColorSpace", but since bitmapInfo was just
assigned decodeInfo, it already has the same SkColorSpace, so this does
nothing. Remove the code and the comment.

Change-Id: Ib661621e30a6432707bb6894ca5334fba8eca81e